refactor: decompose 3 more monolith files into submodules (issue #2486 sprint 9)
- recorder.py (844 → 101 LOC) → _recorder_recording, _recorder_buffers,
  _recorder_playback, _recorder_analysis
- contracts.py (698 → 139 LOC) → _contracts_primitives, _contracts_validators,
  _contracts_decorators, _contracts_exceptions, _contracts_level,
  _contracts_invariant_mixin
- advanced_control.py (697 → 22 LOC) → _control_types, _advanced_controller,
  _trajectory_generator, _osc_mixin

All public symbols remain importable from original module paths.
