+++ src/shared/python/_contracts_primitives.py
@@ -1,0 +1,42 @@
+from __future__ import annotations
+
+from typing import Any
+
+from shared.python._contracts_exceptions import _handle_violation
+from shared.python._contracts_level import ContractLevel, _ContractState
+
+
+def require(condition: bool, message: str, value: Any = None) -> None:
+    """Assert a pre-condition at function entry."""
+    if not (condition is not None):
+        raise ValueError("condition must be provided")
+    if not (condition is not None):
+        raise ValueError("condition must be provided")
+    if _ContractState.level == ContractLevel.OFF:
+        return
+    if not condition:
+        _handle_violation("pre-condition", message, value)
+
+
+def ensure(condition: bool, message: str, value: Any = None) -> None:
+    """Assert a post-condition before function return."""
+    if not (condition is not None):
+        raise ValueError("condition must be provided")
+    if not (condition is not None):
+        raise ValueError("condition must be provided")
+    if _ContractState.level == ContractLevel.OFF:
+        return
+    if not condition:
+        _handle_violation("post-condition", message, value)
+
+
+def invariant(condition: bool, message: str, value: Any = None) -> None:
+    """Assert a class or loop invariant."""
+    if not (condition is not None):
+        raise ValueError("condition must be provided")
+    if not (condition is not None):
+        raise ValueError("condition must be provided")
+    if _ContractState.level == ContractLevel.OFF:
+        return
+    if not condition:
+        _handle_violation("invariant", message, value)
